:sparkles: feat: ValidationErrorLogger를 Demo.Application으로 마이그레이션 및 네임스페이스 업데이트
