+++ NET10/GamePulse/EndPoints/User/Create/CreateEndpointV2.cs
@@ -1,6 +1,6 @@
 using FastEndpoints;
 using Demo.Web.DTO;
-using GamePulse.Processors;
+using Demo.Application.Processors;
 
 namespace GamePulse.EndPoints.User.Create;
