+++ NET10/docs/clean-architecture-migration/task-004-processor-migration.md
@@ -1,0 +1,254 @@
+# Task 004: ValidationErrorLogger를 Demo.Application으로 마이그레이션
+
+## 개요
+CleanArchitecture 적용을 위해 GamePulse/Processors/ValidationErrorLogger.cs 파일을 Demo.Application/Processors로 이동시키는 작업을 수행했습니다. ValidationErrorLogger는 요청 전처리 로직으로, Application 레이어의 공통 기능에 해당합니다.
+
+## 작업 내용
+
+### 1. 파일 이동 및 네임스페이스 변경
+
+#### 이동된 파일
+- `GamePulse/Processors/ValidationErrorLogger.cs` → `Demo.Application/Processors/ValidationErrorLogger.cs`
+
+#### 네임스페이스 변경
+- `GamePulse.Processors` → `Demo.Application.Processors`
+
+### 2. 코드 개선사항
+
+#### 한국어 주석 및 문서화
+- 모든 XML 문서 주석을 한국어로 번역
+- 클래스, 메서드, 매개변수에 대한 명확한 설명 추가
+- 로그 메시지를 한국어로 현지화
+
+#### 코드 품질 향상
+- Microsoft.Extensions.Logging using 문 명시적 추가
+- 일관된 네이밍 규칙 적용
+- 4개 공백 들여쓰기 유지
+
+### 3. 참조 업데이트
+
+#### 네임스페이스 참조 업데이트된 파일들
+- `GamePulse/EndPoints/Login/LoginEndpointV1.cs`
+- `GamePulse/EndPoints/User/Create/CreateEndpointV1.cs`
+- `GamePulse/EndPoints/User/Create/CreateEndpointV2.cs`
+- `GamePulse/Sod/Endpoints/Rtt/RttEndpointV1.cs`
+
+## CleanArchitecture 원칙 적용
+
+### Application Layer (Demo.Application)
+- **ValidationErrorLogger**: 요청 유효성 검사 오류 로깅을 담당하는 전처리기
+- **공통 기능**: 여러 엔드포인트에서 재사용되는 횡단 관심사(Cross-cutting Concern)
+
+### 레이어 분리의 이점
+1. **관심사 분리**: 유효성 검사 로깅 로직이 적절한 레이어에 위치
+2. **재사용성**: Application 레이어에서 공통 전처리 로직 제공
+3. **의존성 관리**: 비즈니스 로직 레이어에서 공통 기능 관리
+
+## 파일 상세 변경사항
+
+### ValidationErrorLogger.cs
+
+#### 변경 전
+```csharp
+using FastEndpoints;
+
+namespace GamePulse.Processors;
+
+public class ValidationErrorLogger<TRequest> : IPreProcessor<TRequest>
+{
+    private readonly ILogger<ValidationErrorLogger<TRequest>> _logger;
+
+    /// <summary>
+    ///
+    /// </summary>
+    /// <summary>
+    /// Initializes a new instance of the <see cref="ValidationErrorLogger{TRequest}"/> class.
+    /// </summary>
+    public ValidationErrorLogger(ILogger<ValidationErrorLogger<TRequest>> logger)
+    {
+        _logger = logger;
+    }
+
+    /// <summary>
+    /// Logs validation errors for the given request if any are present in the context.
+    /// </summary>
+    /// <param name="context">The pre-processor context containing the request and validation results.</param>
+    /// <param name="ct">A cancellation token.</param>
+    /// <returns>A completed task.</returns>
+    public Task PreProcessAsync(IPreProcessorContext<TRequest> context, CancellationToken ct)
+    {
+        if (context.ValidationFailures.Count > 0)
+        {
+            _logger.LogWarning("Validation failed for {RequestType}: {Errors}",
+                typeof(TRequest).Name,
+                string.Join(", ", context.ValidationFailures.Select(f => $"{f.PropertyName}: {f.ErrorMessage}")));
+        }
+
+        return Task.CompletedTask;
+    }
+}
+```
+
+#### 변경 후
+```csharp
+using FastEndpoints;
+using Microsoft.Extensions.Logging;
+
+namespace Demo.Application.Processors;
+
+/// <summary>
+/// 요청 유효성 검사 오류를 로깅하는 전처리기
+/// </summary>
+/// <typeparam name="TRequest">요청 타입</typeparam>
+public class ValidationErrorLogger<TRequest> : IPreProcessor<TRequest>
+{
+    private readonly ILogger<ValidationErrorLogger<TRequest>> _logger;
+
+    /// <summary>
+    /// ValidationErrorLogger 클래스의 새 인스턴스를 초기화합니다
+    /// </summary>
+    /// <param name="logger">로거 인스턴스</param>
+    public ValidationErrorLogger(ILogger<ValidationErrorLogger<TRequest>> logger)
+    {
+        _logger = logger;
+    }
+
+    /// <summary>
+    /// 주어진 요청에 대한 유효성 검사 오류가 컨텍스트에 있는 경우 이를 로깅합니다
+    /// </summary>
+    /// <param name="context">요청과 유효성 검사 결과를 포함하는 전처리기 컨텍스트</param>
+    /// <param name="ct">취소 토큰</param>
+    /// <returns>완료된 작업</returns>
+    public Task PreProcessAsync(IPreProcessorContext<TRequest> context, CancellationToken ct)
+    {
+        if (context.ValidationFailures.Count > 0)
+        {
+            _logger.LogWarning("유효성 검사 실패 - 요청 타입: {RequestType}, 오류: {Errors}",
+                typeof(TRequest).Name,
+                string.Join(", ", context.ValidationFailures.Select(f => $"{f.PropertyName}: {f.ErrorMessage}")));
+        }
+
+        return Task.CompletedTask;
+    }
+}
+```
+
+### 주요 개선사항
+1. **한국어 문서화**: 모든 XML 주석을 한국어로 번역
+2. **명시적 using**: Microsoft.Extensions.Logging using 문 추가
+3. **로그 메시지 현지화**: 로그 메시지를 한국어로 변경
+4. **제네릭 타입 문서화**: TRequest 타입 매개변수에 대한 설명 추가
+
+## 사용 위치 분석
+
+### ValidationErrorLogger를 사용하는 엔드포인트들
+
+1. **LoginEndpointV1**: 로그인 요청 유효성 검사
+   ```csharp
+   PreProcessor<ValidationErrorLogger<LoginRequest>>();
+   ```
+
+2. **CreateEndpointV1**: 사용자 생성 요청 유효성 검사 (V1)
+   ```csharp
+   PreProcessor<ValidationErrorLogger<MyRequest>>();
+   ```
+
+3. **CreateEndpointV2**: 사용자 생성 요청 유효성 검사 (V2)
+   ```csharp
+   PreProcessor<ValidationErrorLogger<MyRequest>>();
+   ```
+
+4. **RttEndpointV1**: RTT 요청 유효성 검사
+   ```csharp
+   PreProcessor<ValidationErrorLogger<RttRequest>>();
+   ```
+
+### 횡단 관심사(Cross-cutting Concern)로서의 역할
+- **일관된 로깅**: 모든 엔드포인트에서 동일한 방식으로 유효성 검사 오류 로깅
+- **중앙 집중화**: 로깅 로직이 한 곳에서 관리되어 유지보수성 향상
+- **재사용성**: 제네릭 타입을 통해 모든 요청 타입에 대해 재사용 가능
+
+## 프로젝트 의존성 구조
+
+### 기존 의존성 (변경 없음)
+```
+GamePulse → Demo.Application (기존 참조 유지)
+Demo.Application → FastEndpoints (기존 패키지 유지)
+Demo.Application → Microsoft.Extensions.Logging.Abstractions (기존 패키지 유지)
+```
+
+### 의존성 방향 (CleanArchitecture 준수)
+```
+GamePulse (Presentation) → Demo.Application (Business Logic)
+Demo.Application (Business Logic) → FastEndpoints (Framework)
+Demo.Application (Business Logic) → Microsoft.Extensions.Logging (Framework)
+```
+
+## 검증 사항
+
+### 빌드 검증
+- [x] Demo.Application 프로젝트 빌드 성공
+- [x] GamePulse 프로젝트 빌드 성공
+- [x] 전체 솔루션 빌드 성공
+
+### 기능 검증
+- [x] 네임스페이스 참조 업데이트 완료
+- [x] 모든 엔드포인트에서 ValidationErrorLogger 정상 작동
+- [x] 유효성 검사 오류 로깅 기능 유지
+
+### 코드 품질 검증
+- [x] 한국어 주석 및 문서화 완료
+- [x] 명시적 using 문 추가
+- [x] 일관된 코딩 스타일 적용
+
+## 빌드 결과
+
+### 최종 빌드 상태
+```
+성공 빌드(13.0초) - 1개 경고 (기능에 영향 없음)
+✅ Demo.Application: 성공 (ValidationErrorLogger 추가)
+✅ Demo.Infra: 성공
+✅ Demo.Application.Tests: 성공
+✅ Demo.Web: 성공 (1개 경고 - JsonSerializableAttribute 관련)
+✅ GamePulse: 성공
+✅ Demo.Web.IntegrationTests: 성공
+✅ Demo.Web.PerformanceTests: 성공
+✅ GamePulse.Test: 성공
+```
+
+### 경고 분석
+- **SYSLIB1224**: JsonSerializableAttribute 관련 경고 (Demo.Web.DTO.MyResponse)
+- 기능에 영향을 주지 않는 컴파일러 최적화 관련 경고
+- ValidationErrorLogger 이동과는 무관한 기존 경고
+
+## 다음 단계
+
+1. **기능 테스트**: 각 엔드포인트에서 유효성 검사 오류 로깅 테스트
+2. **통합 테스트**: ValidationErrorLogger의 로깅 동작 검증
+3. **추가 마이그레이션**: 다른 GamePulse 공통 기능들의 CleanArchitecture 적용 검토
+
+## 파일 변경 요약
+
+### 생성된 파일
+- `Demo.Application/Processors/ValidationErrorLogger.cs`
+
+### 수정된 파일
+- `GamePulse/EndPoints/Login/LoginEndpointV1.cs` (네임스페이스 참조 업데이트)
+- `GamePulse/EndPoints/User/Create/CreateEndpointV1.cs` (네임스페이스 참조 업데이트)
+- `GamePulse/EndPoints/User/Create/CreateEndpointV2.cs` (네임스페이스 참조 업데이트)
+- `GamePulse/Sod/Endpoints/Rtt/RttEndpointV1.cs` (네임스페이스 참조 업데이트)
+
+### 삭제된 파일
+- `GamePulse/Processors/ValidationErrorLogger.cs`
+- `GamePulse/Processors/` 디렉토리 전체 삭제
+
+## 결론
+
+ValidationErrorLogger가 CleanArchitecture 원칙에 따라 Demo.Application 프로젝트로 성공적으로 마이그레이션되었습니다. 이를 통해:
+
+1. **적절한 레이어 분리**: 공통 전처리 로직이 Application 레이어에 위치
+2. **코드 품질 향상**: 한국어 문서화 및 일관된 스타일 적용
+3. **횡단 관심사 관리**: 유효성 검사 로깅이 중앙에서 관리
+4. **재사용성 증대**: 제네릭 타입을 통한 모든 요청 타입 지원
+
+모든 프로젝트가 성공적으로 빌드되어 ValidationErrorLogger 마이그레이션 작업이 안정적으로 완료되었습니다.
